Refactored file formats into stctrl/src/file
