+++ components/stctrl/src/file/mod.rs
@@ -1,0 +1,2 @@
+pub mod invoice;
+pub mod receipt;
