+++ components/stctrl/src/lib.rs
@@ -22,5 +22,4 @@ pub mod config;
 pub mod funds;
 pub mod info;
 
-pub mod invoice;
-pub mod receipt;
+pub mod file;
